+++ kong/partial.go
@@ -7,7 +7,7 @@ type Partial struct {
 	Name      *string       `json:"name,omitempty" yaml:"name,omitempty"`
 	Type      *string       `json:"type,omitempty" yaml:"type,omitempty"`
 	Config    Configuration `json:"config,omitempty" yaml:"config,omitempty"`
-	CreatedAt *int          `json:"created_at,omitempty" yaml:"created_at,omitempty"`
+	CreatedAt *float64          `json:"created_at,omitempty" yaml:"created_at,omitempty"`
 	UpdatedAt *int          `json:"updated_at,omitempty" yaml:"updated_at,omitempty"`
 	Tags      []*string     `json:"tags,omitempty" yaml:"tags,omitempty"`
 }
